refactor(utils): consolidate remark style helpers
